Changes to allow pip install
Changes to files and structure to make pip installable
